Move create_test_datatree fixture to conftest.py
