+++ datatree/tests/test_io.py
@@ -3,7 +3,6 @@ import pytest
 from datatree.io import open_datatree
 from datatree.testing import assert_equal
 from datatree.tests import requires_h5netcdf, requires_netCDF4, requires_zarr
-from datatree.tests.test_datatree import create_test_datatree, simple_datatree
 
 
 class TestIO:
